+++ src/components/AaravViewerWrapper.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { AaravViewer } from '../core/AaravViewer';
 
 type AaravViewerWrapperProps = {
-  aaravMainViewer: AaravViewer;
+  aaravViewer: AaravViewer;
 };
 
 /**
@@ -10,9 +10,7 @@ type AaravViewerWrapperProps = {
  */
 class AaravViewerWrapper extends React.Component<AaravViewerWrapperProps> {
   componentWillUnmount() {
-    const { aaravMainViewer } = this.props;
-
-    const aaravViewer = aaravMainViewer;
+    const { aaravViewer } = this.props;
 
     if (aaravViewer.attached) {
       aaravViewer.detach();
@@ -23,9 +21,7 @@ class AaravViewerWrapper extends React.Component<AaravViewerWrapperProps> {
    * @argument {HTMLDivElement} container
    */
   containerRef = (container: HTMLDivElement | null) => {
-    const { aaravMainViewer } = this.props;
-
-    const aaravViewer = aaravMainViewer;
+    const { aaravViewer } = this.props;
 
     if (aaravViewer.attached) {
       aaravViewer.detach();